fix: Замечания спринт 10 ревью 1
