+++ src/main/java/ru/yandex/practicum/filmorate/exeptions/FilmIdUnknownException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exeptions;
 
 public class FilmIdUnknownException extends RuntimeException {
-    public FilmIdUnknownException(final long filmId) {
-        super("Фильм с id: " + filmId +" не найден");
+    public FilmIdUnknownException(final String message) {
+        super(message);
     }
 }
