+++ src/main/java/ru/yandex/practicum/filmorate/exeptions/MethodNotImplementedException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exeptions;
 
 public class MethodNotImplementedException extends RuntimeException {
-    public MethodNotImplementedException(final String method) {
-        super("Метод \"" + method + "\" ещё не реализован");
+    public MethodNotImplementedException(final String message) {
+        super(message);
     }
 }
